Cover components (fix #305)
