+++ src/js/constants/AppConstants.js
@@ -31,7 +31,6 @@ export default keymirror({
     APP_LIBRARY_REFRESH_END      : null,
     APP_LIBRARY_REFRESH_PROGRESS : null,
     APP_LIBRARY_SET_TRACKSCURSOR : null,
-    APP_LIBRARY_FETCHED_COVER    : null,
     APP_LIBRARY_REMOVE_TRACKS    : null,
 
     APP_PLAYLISTS_REFRESH  : null,
