+++ src/js/reducers/library.js
@@ -28,13 +28,6 @@ export default (state = {}, payload) => {
             return newState;
         }
 
-        case(AppConstants.APP_LIBRARY_FETCHED_COVER): {
-            return {
-                ...state,
-                cover: payload.cover || null
-            };
-        }
-
         case(AppConstants.APP_LIBRARY_ADD_FOLDERS): {
             const folders    = payload.folders;
             let musicFolders = app.config.get('musicFolders');
